requirements.txt
requests
beautifulsoup4
utils
pymongo
vk-api-boroda34